tests(pix-editor): extract pdf export into dedicated service to allow testing it
